:recycle: Simplifying cli entrypoint
