+++ Sources/SwaggerSwiftCore/Models/HTTPMethod.swift
@@ -1,0 +1,3 @@
+enum HTTPMethod: String {
+    case get, put, post, delete, patch, options, head
+}
