+++ Sources/SwaggerSwiftCore/Static Files/DummyTestFile.swift
@@ -1,0 +1,9 @@
+let dummyTest = """
+import XCTest
+
+public struct DummyTest {
+    func testNetwork() {
+        XCTAssertTrue(true)
+    }
+}
+"""
