+++ Sources/SwaggerSwiftCore/Static Files/JsonParsingErrorExtensionFile.swift
@@ -1,0 +1,7 @@
+let jsonParsingErrorExtension = """
+import Foundation
+
+public enum JSONParsingError: Error {
+    case invalidDate(String)
+}
+"""
