+++ Sources/SwaggerSwiftCore/Static Files/NetworkInterceptorProtocolFile.swift
@@ -1,0 +1,8 @@
+let networkInterceptor = """
+import Foundation
+
+public protocol NetworkInterceptor {
+    func networkWillPerformRequest(_ request: URLRequest) -> URLRequest
+    func networkDidPerformRequest(urlRequest: URLRequest, urlResponse: URLResponse?, data: Data?, error: Error?) -> Bool
+}
+"""
