+++ Sources/SwaggerSwiftCore/Static Files/URLQueryExtensionFile.swift
@@ -1,0 +1,25 @@
+let urlQueryItemExtension = """
+import Foundation
+
+public extension URLQueryItem {
+    init(name: String, value: Bool) {
+        self.init(name: name, value: value ? "true" : "false")
+    }
+
+    init(name: String, value: Int) {
+        self.init(name: name, value: String(value))
+    }
+
+    init(name: String, value: Int64) {
+        self.init(name: name, value: String(value))
+    }
+
+    init(name: String, value: Double) {
+        self.init(name: name, value: String(value))
+    }
+
+    init(name: String, value: Date) {
+        self.init(name: name, value: ISO8601DateFormatter().string(from: value))
+    }
+}
+"""
