--- Sources/SwaggerSwiftCore/SwaggerFileParser/Models/Service.swift
@@ -1,4 +1,0 @@
-struct Service: Decodable {
-    let branch: String?
-    let path: String?
-}
